Issue #32 - Copy responses from source question
* Use long for Survey Group IDs
* Prefill: Use previous instances of a survey, if exists.
* Otherwise, use the instance of the registration survey, if exists

+++ survey/src/com/gallatinsystems/survey/device/activity/RecordListActivity.java
@@ -203,7 +203,7 @@ public class RecordListActivity extends ActionBarActivity implements
         public Fragment getItem(int position) {
             Fragment fragment = null;
             Bundle extras = new Bundle();
-            extras.putInt(EXTRA_SURVEY_GROUP_ID, mSurveyGroup.getId());
+            extras.putLong(EXTRA_SURVEY_GROUP_ID, mSurveyGroup.getId());
             switch (position) {
                 case POSITION_LIST:
                     fragment =  new SurveyedLocaleListFragment();
