+++ survey/src/com/gallatinsystems/survey/device/async/loader/SurveyListLoader.java
@@ -23,9 +23,9 @@ import com.gallatinsystems.survey.device.async.loader.base.DataLoader;
 import com.gallatinsystems.survey.device.dao.SurveyDbAdapter;
 
 public class SurveyListLoader extends DataLoader<Cursor> {
-    private int mSurveyGroupId;
+    private long mSurveyGroupId;
     
-    public SurveyListLoader(Context context, SurveyDbAdapter db, int surveyGroupId) {
+    public SurveyListLoader(Context context, SurveyDbAdapter db, long surveyGroupId) {
         super(context, db);
         mSurveyGroupId = surveyGroupId;
     }
