+++ survey/src/com/gallatinsystems/survey/device/domain/SurveyedLocale.java
@@ -27,12 +27,12 @@ public class SurveyedLocale implements Serializable {
     
     private String mId;
     private String mName;
-    private int mSurveyGroupId;
+    private long mSurveyGroupId;
     private double mLatitude;
     private double mLongitude;
     private List<SurveyInstance> mSurveyInstances = null;
 
-    public SurveyedLocale (String id, String name, int surveyGroupId, double latitude, double longitude) {
+    public SurveyedLocale (String id, String name, long surveyGroupId, double latitude, double longitude) {
         mId = id;
         mName = name;
         mSurveyGroupId = surveyGroupId;
@@ -40,7 +40,7 @@ public class SurveyedLocale implements Serializable {
         mLongitude = longitude;
     }
     
-    public int getSurveyGroupId() {
+    public long getSurveyGroupId() {
         return mSurveyGroupId;
     }
